feat(create): give /create its own app-shell top bar
