+++ www/src/routes/_app/create.tsx
@@ -12,7 +12,9 @@ import {
   saveStoredPreset,
 } from '@/modules/create/preset/storage'
 import { PreviewPanel } from '@/modules/create/preview/preview-panel'
+import { StudioChromeProvider } from '@/modules/create/studio/chrome'
 import { StudioPanel } from '@/modules/create/studio/studio-panel'
+import { CreateTopBar } from '@/modules/create/studio/top-bar'
 
 type MobilePane = 'customize' | 'preview'
 
@@ -69,33 +71,43 @@ function CreatePage() {
     saveStoredPreset(designSystem)
   }, [designSystem])
 
-  // Opt-in exploration of the redesigned control panel + floating panel lab.
-  if (lab) return <LabExperience />
-
+  // The builder owns its top bar: the global site Header is suppressed on /create
+  // (see _app/route.tsx), and CreateTopBar takes its place — same height token,
+  // sticky, Logo / GitHub / ThemeToggle — carrying the builder's primary actions
+  // (depth switch, surprise-me, undo, reset) from StudioChromeProvider.
   return (
-    <div className="flex h-[calc(100svh-var(--header-height))] min-h-0 flex-1 flex-col gap-3 p-4 pt-2 lg:flex-row lg:gap-6 lg:p-6 lg:pt-2">
-      {/* Mobile-only view switcher — hidden once the two panes fit side by side. */}
-      <ToggleButtonGroup
-        aria-label="Editor view"
-        selectionMode="single"
-        disallowEmptySelection
-        size="sm"
-        selectedKeys={[mobilePane]}
-        onSelectionChange={(keys) => {
-          const next = keys.values().next().value
-          if (next === 'customize' || next === 'preview') setMobilePane(next)
-        }}
-        className="w-full shrink-0 *:flex-1 lg:hidden"
-      >
-        <ToggleButton id="customize">Customize</ToggleButton>
-        <ToggleButton id="preview">Preview</ToggleButton>
-      </ToggleButtonGroup>
-      <StudioPanel
-        className={cn(mobilePane === 'preview' && 'max-lg:hidden')}
-      />
-      <PreviewPanel
-        className={cn(mobilePane === 'customize' && 'max-lg:hidden')}
-      />
-    </div>
+    <StudioChromeProvider>
+      <CreateTopBar />
+      {lab ? (
+        // Opt-in exploration of the redesigned control panel + floating panel lab.
+        <LabExperience />
+      ) : (
+        <div className="flex h-[calc(100svh-var(--header-height))] min-h-0 flex-1 flex-col gap-3 p-4 pt-2 lg:flex-row lg:gap-6 lg:p-6 lg:pt-2">
+          {/* Mobile-only view switcher — hidden once the two panes fit side by side. */}
+          <ToggleButtonGroup
+            aria-label="Editor view"
+            selectionMode="single"
+            disallowEmptySelection
+            size="sm"
+            selectedKeys={[mobilePane]}
+            onSelectionChange={(keys) => {
+              const next = keys.values().next().value
+              if (next === 'customize' || next === 'preview')
+                setMobilePane(next)
+            }}
+            className="w-full shrink-0 *:flex-1 lg:hidden"
+          >
+            <ToggleButton id="customize">Customize</ToggleButton>
+            <ToggleButton id="preview">Preview</ToggleButton>
+          </ToggleButtonGroup>
+          <StudioPanel
+            className={cn(mobilePane === 'preview' && 'max-lg:hidden')}
+          />
+          <PreviewPanel
+            className={cn(mobilePane === 'customize' && 'max-lg:hidden')}
+          />
+        </div>
+      )}
+    </StudioChromeProvider>
   )
 }
